SOAP/requirements.txt
flask
lxml
psycopg2-binary
zeep
